have product's ID's, need to present them in profile_page. everything works, merge to dev
