+++ lib/core/models/givit_user.dart
@@ -7,6 +7,7 @@ class GivitUser {
   final String fullName;
   final int phoneNumber;
   final String role;
+  List<String> products;
 
   GivitUser(
       {this.email = '',
@@ -14,7 +15,8 @@ class GivitUser {
       this.fullName = '',
       this.phoneNumber = 0,
       this.role = '',
-      this.uid = ''});
+      this.uid = '',
+      this.products});
 
   factory GivitUser.fromFirestorUser(DocumentSnapshot userSnapshout) {
     return GivitUser(
@@ -24,6 +26,7 @@ class GivitUser {
       fullName: userSnapshout['Full Name'],
       phoneNumber: userSnapshout['Phone Number'],
       role: userSnapshout['Role'],
+      products: List.from(userSnapshout['Products']),
     );
   }
 }
